Adaptive-Barycenters_v2/requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
absl-py=0.13.0=pypi_0
astor=0.8.0=py37_0
astunparse=1.6.3=pypi_0
blas=1.0=mkl
blinker=1.4=py37_0
boost-cpp=1.70.0=ha2d47e9_1
bzip2=1.0.8=h7b6447c_0
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2021.5.30=ha878542_0
cached-property=1.5.2=pypi_0
cachetools=3.1.1=py_0
certifi=2021.5.30=py37h89c1867_0
cffi=1.14.0=py37h2e261b9_0
cgal-cpp=5.0.1=h261e690_0
chardet=3.0.4=py37_1003
click=7.1.2=py_0
cloudpickle=1.4.1=py_0
cryptography=2.9.2=py37h1ba5d50_0
cudatoolkit=11.1.74=h6bb024c_0
cycler=0.10.0=py37_0
cytoolz=0.10.1=py37h7b6447c_0
dask-core=2.16.0=py_0
dbus=1.13.14=hb2f20db_0
decorator=4.4.2=py_0
dm-tree=0.1.6=pypi_0
eigen=3.3.8=h0efe328_0
et_xmlfile=1.0.1=py_1001
expat=2.2.6=he6710b0_0
ffmpeg=4.2.2=h20bf706_0
flatbuffers=1.12=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
gast=0.4.0=pypi_0
glib=2.63.1=h5a9c865_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=1.14.1=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
grpcio=1.34.1=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
gudhi=3.1.1=py37hb3f55d8_0
h5py=3.1.0=pypi_0
hdf5=1.10.4=hb1b8bf9_0
icu=58.2=he6710b0_3
idna=2.9=py_1
imagecodecs-lite=2019.12.3=py37h8f50634_0
imageio=2.8.0=py_0
intel-openmp=2020.1=217
jdcal=1.4.1=py_0
joblib=0.14.1=py_0
jpeg=9b=h024ee3a_2
keras-applications=1.0.8=py_0
keras-base=2.3.1=py37_0
keras-nightly=2.5.0.dev2021032900=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.2.0=py37hfd86e86_0
lame=3.100=h7b6447c_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libidn2=2.3.1=h27cfd23_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.11.4=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
lmdb=0.98=pypi_0
markdown=3.1.1=py37_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37hef1b27d_0
mkl=2020.1=217
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
mpfr=4.0.2=he80fd80_1
ncurses=6.2=he6710b0_1
nettle=3.7.2=hbbd107a_1
networkx=2.4=py_0
ninja=1.9.0=py37hfd86e86_0
numpy=1.19.5=pypi_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
openh264=2.1.0=hd408876_0
openpyxl=3.0.5=py_0
openssl=1.1.1h=h516909a_0
opt-einsum=3.3.0=pypi_0
pcre=8.43=he6710b0_0
pillow=7.1.2=py37hb39fc2d_0
pip=20.0.2=py37_3
protobuf=3.11.4=py37he6710b0_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=py_0
pyjwt=1.7.1=py37_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h05f1152_2
pysocks=1.7.1=py37_0
python=3.7.7=hcf32534_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.9.0.dev20210527=py3.7_cuda11.1_cudnn8.0.5_0
pywavelets=1.1.1=py37h7b6447c_0
pyyaml=5.3.1=py37h7b6447c_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
requests=2.23.0=py37_0
requests-oauthlib=1.3.0=py_0
rsa=4.0=py_0
scikit-image=0.17.2=py37h0da4684_0
scikit-learn=0.22.1=py37hd81dba3_0
scipy=1.4.1=py37h0b6359f_0
setuptools=46.2.0=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=pypi_0
sqlite=3.31.1=h62c20be_1
tensorboard=2.6.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.5.0=pypi_0
tensorflow-estimator=2.5.0=pypi_0
tensorflow-gan=2.0.0=pypi_0
tensorflow-hub=0.12.0=pypi_0
tensorflow-probability=0.13.0=pypi_0
termcolor=1.1.0=py37_1
tifffile=2020.5.30=py_0
tk=8.6.8=hbc83047_0
toolz=0.10.0=py_0
torchaudio=0.9.0.dev20210527=py37
torchvision=0.11.0.dev20210527=py37_cu111
tornado=6.0.4=py37h7b6447c_1
typing=3.10.0.0=py37h06a4308_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.25.8=py37_0
werkzeug=0.16.1=py_0
wheel=0.37.0=pypi_0
wrapt=1.12.1=py37h7b6447c_1
x264=1!157.20191217=h7b6447c_0
xlsxwriter=1.3.7=py_0
xz=5.2.5=h7b6447c_0
yaml=0.1.7=had09818_2
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
